A research manusrcipt on the hydrodynamic model
The project was funded by Ministry of Science and Technology, Bangladesh. The study is modifying, and I hope to submit it to a publisher very soon including me as a 1st author (actually the research was conducted by me, although Principal Investigator was aware about every step and joined the regular meeting of Ministry with me).